# development-plan/openapi.yaml
openapi: 3.0.3
info:
  title: Daily Ledger API
  version: 0.1.0-draft
  description: |
    Daily Ledger REST API draft.
    This specification follows the current planning documents and is intended
    as an implementation baseline for FastAPI.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
  /api/transactions:
    post:
      tags: [Transactions]
      summary: 新增交易
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '201':
          description: 建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Transactions]
      summary: 查詢交易清單
      operationId: listTransactions
      parameters:
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/type'
        - $ref: '#/components/parameters/category_main'
        - $ref: '#/components/parameters/category_sub'
        - $ref: '#/components/parameters/keyword'
        - $ref: '#/components/parameters/amount_min'
        - $ref: '#/components/parameters/amount_max'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/transactions/{id}:
    get:
      tags: [Transactions]
      summary: 取得單筆交易
      operationId: getTransaction
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Transactions]
      summary: 全欄位覆蓋更新交易
      operationId: updateTransaction
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Transactions]
      summary: 刪除交易
      operationId: deleteTransaction
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 刪除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/transactions/date_range:
    get:
      tags: [Transactions]
      summary: 取得資料日期範圍與筆數
      operationId: getTransactionsDateRange
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DateRangeResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/categories:
    get:
      tags: [Categories]
      summary: 取得分類清單（平鋪陣列）
      operationId: listCategories
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [E, I]
          description: 類型篩選
        - name: include_count
          in: query
          required: false
          schema:
            type: integer
            enum: [0, 1]
            default: 0
          description: 1 時附加分類引用筆數
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/CategoryItem'
                    - $ref: '#/components/schemas/CategoryItemWithCount'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Categories]
      summary: 新增分類
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: 建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryItem'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Categories]
      summary: 刪除分類
      operationId: deleteCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDeleteRequest'
      responses:
        '200':
          description: 刪除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/categories/merge:
    post:
      tags: [Categories]
      summary: 合併分類（含改名情境）
      operationId: mergeCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryMergeRequest'
      responses:
        '200':
          description: 合併成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryMergeResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/import/preview:
    post:
      tags: [ImportExport]
      summary: 匯入預覽（不上線寫入）
      operationId: importPreview
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 預覽成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreviewResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/import/confirm:
    post:
      tags: [ImportExport]
      summary: 匯入確認（實際寫入）
      operationId: importConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportConfirmRequest'
      responses:
        '200':
          description: 寫入成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportConfirmResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/export:
    get:
      tags: [ImportExport]
      summary: 匯出交易 CSV
      operationId: exportTransactions
      parameters:
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
      responses:
        '200':
          description: 匯出成功
          headers:
            Content-Disposition:
              description: 下載檔名
              schema:
                type: string
                example: attachment; filename=daily_ledger_20260101_20260331.csv
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/report/monthly:
    get:
      tags: [Reports]
      summary: 月度收支摘要
      operationId: reportMonthly
      parameters:
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/report_type'
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportMonthlyItem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/report/category:
    get:
      tags: [Reports]
      summary: 分類佔比
      operationId: reportCategory
      parameters:
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/report_type'
        - name: level
          in: query
          required: false
          schema:
            type: string
            enum: [main, sub]
            default: main
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportCategoryItem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/report/trend:
    get:
      tags: [Reports]
      summary: 月度趨勢
      operationId: reportTrend
      parameters:
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/report_type'
        - $ref: '#/components/parameters/category_main'
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportTrendItem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 交易 ID（UUID4 前 8 碼）
    from:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date
      description: 起始日期（YYYY-MM-DD）
    to:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date
      description: 結束日期（YYYY-MM-DD）
    type:
      name: type
      in: query
      required: false
      schema:
        type: string
        enum: [E, I, all]
      description: 交易類型篩選
    report_type:
      name: type
      in: query
      required: false
      schema:
        type: string
        enum: [E, I, all]
        default: all
      description: 報表類型篩選
    category_main:
      name: category_main
      in: query
      required: false
      schema:
        type: string
    category_sub:
      name: category_sub
      in: query
      required: false
      schema:
        type: string
    keyword:
      name: keyword
      in: query
      required: false
      schema:
        type: string
      description: 明細或備註關鍵字
    amount_min:
      name: amount_min
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
    amount_max:
      name: amount_max
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
  responses:
    BadRequestError:
      description: 請求格式錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: 資料衝突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: 參數驗證失敗
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    TransactionCreate:
      allOf:
        - $ref: '#/components/schemas/TransactionBase'
    TransactionUpdate:
      allOf:
        - $ref: '#/components/schemas/TransactionBase'
    Transaction:
      allOf:
        - $ref: '#/components/schemas/TransactionBase'
        - type: object
          required: [id, 建立時間]
          properties:
            id:
              type: string
              example: a3f7c2d1
            建立時間:
              type: string
              format: date-time
              example: 2026-04-07T13:20:30+08:00
...
    TransactionListResponse:
      type: object
      required: [items, total, page, size, pages, summary]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        total:
          type: integer
          example: 975
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 100
        pages:
          type: integer
          example: 10
        summary:
          $ref: '#/components/schemas/Summary'
    DateRangeResponse:
      type: object
      required: [min, max, count]
      properties:
        min:
          type: string
          format: date
          nullable: true
          example: 2025-01-01
        max:
          type: string
          format: date
          nullable: true
          example: 2026-04-07
        count:
          type: integer
          example: 975
    CategoryItem:
      type: object
      required: [類型, 主類, 次類]
      properties:
        類型:
          type: string
          enum: [E, I]
          example: E
        主類:
          type: string
          example: 餐飲費
        次類:
          type: string
          example: 早餐
    CategoryItemWithCount:
      allOf:
        - $ref: '#/components/schemas/CategoryItem'
        - type: object
          required: [count]
          properties:
            count:
              type: integer
              example: 234
    CategoryCreateRequest:
      $ref: '#/components/schemas/CategoryItem'
    CategoryDeleteRequest:
      $ref: '#/components/schemas/CategoryItem'
    CategoryMergeRequest:
      type: object
      required: [source, target]
      properties:
        source:
          $ref: '#/components/schemas/CategoryItem'
        target:
          $ref: '#/components/schemas/CategoryItem'
    CategoryMergeResponse:
      type: object
      required: [updated_count, removed_category]
      properties:
        updated_count:
          type: integer
          example: 56
        removed_category:
          $ref: '#/components/schemas/CategoryItem'
...
    ImportPreviewResponse:
      type: object
      required: [preview_token, stats, new_categories, sample_transactions]
      properties:
        preview_token:
          type: string
          example: 2f9a6f9b9e8a4e3bb5d2b7df31d00123
        stats:
          $ref: '#/components/schemas/ImportPreviewStats'
        new_categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryItem'
        sample_transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    ImportConfirmRequest:
      type: object
      required: [preview_token]
      properties:
        preview_token:
          type: string
    ImportConfirmResponse:
      type: object
      required: [inserted_count, skipped_count, added_categories]
      properties:
        inserted_count:
          type: integer
          example: 450
        skipped_count:
          type: integer
          example: 520
        added_categories:
          type: integer
          example: 3
    ReportMonthlyItem:
      type: object
      required: [month, income, expense, net]
      properties:
        month:
          type: string
          example: 2026-04
        income:
          type: integer
          example: 35000
        expense:
          type: integer
          example: 21000
        net:
          type: integer
          example: 14000
    ReportCategoryItem:
      type: object
      required: [類型, 主類, 次類, amount, ratio]
      properties:
        類型:
          type: string
          enum: [E, I]
          example: E
        主類:
          type: string
          example: 餐飲費
        次類:
          type: string
          example: 午餐
        amount:
          type: integer
          example: 2500
        ratio:
          type: number
          format: float
          example: 0.18
    ReportTrendItem:
      type: object
      required: [month, income, expense, net]
      properties:
        month:
          type: string
          example: 2026-04
        income:
          type: integer
          example: 35000
        expense:
          type: integer
          example: 21000
        net:
          type: integer
          example: 14000
    SuccessMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: ok
...